protect against phishing attacks
to prevent google ads from being abused on our main website domain,
add a special route which triggers only on google ads bot.

The idea is that while the google check person is verifying the ad, they
should see the final url preview and realize that the site is a scam and
reject it.